# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a DDL Guard server API.",
        "title": "DDL Guard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/final-grades/{final_grade_uuid}/homework-scores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定期末成绩记录下的所有作业成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeworkScore"
                ],
                "summary": "获取作业成绩列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "final_grade_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.HomeworkScoreListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定期末成绩记录添加作业成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeworkScore"
                ],
                "summary": "创建作业成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "final_grade_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Homework Score Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.CreateHomeworkScoreReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.CreateHomeworkScoreResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/final-grades/{final_grade_uuid}/quiz-scores": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定期末成绩记录下的所有小测成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuizScore"
                ],
                "summary": "获取小测成绩列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "final_grade_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.QuizScoreListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定期末成绩记录添加小测成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuizScore"
                ],
                "summary": "创建小测成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Final Grade UUID",
                        "name": "final_grade_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Quiz Score Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.CreateQuizScoreReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.CreateQuizScoreResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/homework-scores/{uuid}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新作业成绩的信息，自动重算最终成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeworkScore"
                ],
                "summary": "更新作业成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Homework Score UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Homework Score Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdateHomeworkScoreReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.UpdateHomeworkScoreResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除作业成绩，自动重算最终成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HomeworkScore"
                ],
                "summary": "删除作业成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Homework Score UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/quiz-scores/{uuid}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新小测成绩的信息，自动重算最终成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuizScore"
                ],
                "summary": "更新小测成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Score UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Quiz Score Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdateQuizScoreReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.UpdateQuizScoreResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除小测成绩，自动重算最终成绩",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuizScore"
                ],
                "summary": "删除小测成绩",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Score UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "schema.CreateHomeworkScoreReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                }
            }
        },
        "schema.CreateHomeworkScoreResp": {
            "type": "object",
            "properties": {
                "final_grade_id": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                },
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.CreateQuizScoreReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                }
            }
        },
        "schema.CreateQuizScoreResp": {
            "type": "object",
            "properties": {
                "final_grade_id": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                },
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
...
        "schema.HomeworkScoreListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-20T10:00:00+08:00"
                },
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.HomeworkScoreListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.HomeworkScoreListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "schema.QuizScoreListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-20T10:00:00+08:00"
                },
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.QuizScoreListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.QuizScoreListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "schema.UpdateHomeworkScoreReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                }
            }
        },
        "schema.UpdateHomeworkScoreResp": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "作业1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        },
        "schema.UpdateQuizScoreReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                }
            }
        },
        "schema.UpdateQuizScoreResp": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "小测1"
                },
                "score": {
                    "type": "number",
                    "example": 90
                },
                "uuid": {
                    "type": "string",
                    "example": "7a178766-4b8e-4e99-ab4c-843f7dbd95fd"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}